+++ projects/eval-adxl355-pmdz/builds.json
@@ -1,100 +1,100 @@
 {
   "stm32": {
     "dummy_example": {
-      "flags" : "DUMMY_EXAMPLE=y IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=n IIO_LWIP_EXAMPLE=n ADXL355=y HARDWARE=sdp-ck1z.ioc"
+      "flags" : "EXAMPLE=dummy ADXL355=y HARDWARE=sdp-ck1z.ioc"
     },
     "iio_example": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=y IIO_TRIGGER_EXAMPLE=n IIO_LWIP_EXAMPLE=n ADXL355=y HARDWARE=sdp-ck1z.ioc"
+      "flags" : "EXAMPLE=iio_example ADXL355=y HARDWARE=sdp-ck1z.ioc"
     },
     "iio_trigger_example": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=y IIO_LWIP_EXAMPLE=n ADXL355=y HARDWARE=sdp-ck1z.ioc"
+      "flags" : "EXAMPLE=iio_trigger_example ADXL355=y HARDWARE=sdp-ck1z.ioc"
     },
     "iio_lwip_example": {
-	"flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=n IIO_LWIP_EXAMPLE=y ADXL355=y HARDWARE=adin1110ebz.ioc"
+	"flags" : "EXAMPLE=iio_lwip_example ADXL355=y HARDWARE=adin1110ebz.ioc"
     }
   },
   "maxim": {
     "dummy_example_max32650_adxl355": {
-      "flags" : "DUMMY_EXAMPLE=y IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=n TARGET=max32650 ADXL355=y"
+      "flags" : "EXAMPLE=dummy TARGET=max32650 ADXL355=y"
     },
     "iio_example_max32650_adxl355": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=y IIO_TRIGGER_EXAMPLE=n TARGET=max32650 ADXL355=y"
+      "flags" : "EXAMPLE=iio_example TARGET=max32650 ADXL355=y"
     },
     "dummy_example_max32655_adxl355": {
-      "flags" : "DUMMY_EXAMPLE=y IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=n TARGET=max32655 ADXL355=y"
+      "flags" : "EXAMPLE=dummy TARGET=max32655 ADXL355=y"
     },
     "iio_example_max32655_adxl355": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=y IIO_TRIGGER_EXAMPLE=n TARGET=max32655 ADXL355=y"
+      "flags" : "EXAMPLE=iio_example TARGET=max32655 ADXL355=y"
     },
     "iio_trigger_example_max32655_adxl355": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=y TARGET=max32655 ADXL355=y"
+      "flags" : "EXAMPLE=iio_trigger_example TARGET=max32655 ADXL355=y"
     },
     "dummy_example_max32660_adxl355": {
-      "flags" : "DUMMY_EXAMPLE=y IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=n TARGET=max32660 ADXL355=y"
+      "flags" : "EXAMPLE=dummy TARGET=max32660 ADXL355=y"
     },
     "iio_example_max32660_adxl355": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=y IIO_TRIGGER_EXAMPLE=n TARGET=max32660 ADXL355=y"
+      "flags" : "EXAMPLE=iio_example TARGET=max32660 ADXL355=y"
     },
     "dummy_example_max32665_adxl355": {
-      "flags" : "DUMMY_EXAMPLE=y IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=n TARGET=max32665 ADXL355=y"
+      "flags" : "EXAMPLE=dummy TARGET=max32665 ADXL355=y"
     },
     "iio_example_max32665_adxl355": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=y IIO_TRIGGER_EXAMPLE=n TARGET=max32665 ADXL355=y"
+      "flags" : "EXAMPLE=iio_example TARGET=max32665 ADXL355=y"
     },
     "dummy_example_max78000_adxl355": {
-      "flags" : "DUMMY_EXAMPLE=y IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=n TARGET=max78000 ADXL355=y"
+      "flags" : "EXAMPLE=dummy TARGET=max78000 ADXL355=y"
     },
     "iio_example_max78000_adxl355": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=y IIO_TRIGGER_EXAMPLE=n TARGET=max78000 ADXL355=y"
+      "flags" : "EXAMPLE=iio_example TARGET=max78000 ADXL355=y"
     },
     "dummy_example_max78000_adxl359": {
-      "flags" : "DUMMY_EXAMPLE=y IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=n TARGET=max78000 ADXL359=y"
+      "flags" : "EXAMPLE=dummy TARGET=max78000 ADXL359=y"
     },
     "iio_example_max78000_adxl359": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=y IIO_TRIGGER_EXAMPLE=n TARGET=max78000 ADXL359=y"
+      "flags" : "EXAMPLE=iio_example TARGET=max78000 ADXL359=y"
     },
     "iio_trigger_example_max78000_adxl359": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=y TARGET=max32655 ADXL359=y"
+      "flags" : "EXAMPLE=iio_trigger_example TARGET=max32655 ADXL359=y"
     },
     "dummy_example_max32655_adxl357": {
-      "flags" : "DUMMY_EXAMPLE=y IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=n TARGET=max32655 ADXL357=y"
+      "flags" : "EXAMPLE=dummy TARGET=max32655 ADXL357=y"
     },
     "iio_example_max32655_adxl357": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=y IIO_TRIGGER_EXAMPLE=n TARGET=max32655 ADXL357=y"
+      "flags" : "EXAMPLE=iio_example TARGET=max32655 ADXL357=y"
     },
     "iio_trigger_example_max32655_adxl357": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=y TARGET=max32655 ADXL357=y"
+      "flags" : "EXAMPLE=iio_trigger_example TARGET=max32655 ADXL357=y"
     },
     "dummy_example_max32690_adxl355": {
-      "flags" : "DUMMY_EXAMPLE=y IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=n TARGET=max32690 ADXL355=y"
+      "flags" : "EXAMPLE=dummy TARGET=max32690 ADXL355=y"
     },
     "iio_example_max32690_adxl355": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=y IIO_TRIGGER_EXAMPLE=n TARGET=max32690 ADXL355=y"
+      "flags" : "EXAMPLE=iio_example TARGET=max32690 ADXL355=y"
     },
     "iio_trigger_example_max32690_adxl355": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=y TARGET=max32690 ADXL355=y"
+      "flags" : "EXAMPLE=iio_trigger_example TARGET=max32690 ADXL355=y"
     }
   },
   "pico": {
     "iio_example": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=y IIO_TRIGGER_EXAMPLE=n ADXL355=y"
+      "flags" : "EXAMPLE=iio_example ADXL355=y"
     },
     "iio_trigger_example": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=y ADXL355=y"
+      "flags" : "EXAMPLE=iio_trigger_example ADXL355=y"
     },
     "dummy_example": {
-      "flags" : "DUMMY_EXAMPLE=y IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=n ADXL355=y"
+      "flags" : "EXAMPLE=dummy ADXL355=y"
     }
   },
   "aducm3029": {
     "dummy_example": {
-      "flags" : "DUMMY_EXAMPLE=y IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=n ADXL355=y"
+      "flags" : "EXAMPLE=dummy ADXL355=y"
     },
     "iio_example": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=y IIO_TRIGGER_EXAMPLE=n ADXL355=y"
+      "flags" : "EXAMPLE=iio_example ADXL355=y"
     },
     "iio_trigger_example": {
-      "flags" : "DUMMY_EXAMPLE=n IIO_EXAMPLE=n IIO_TRIGGER_EXAMPLE=y ADXL355=y"
+      "flags" : "EXAMPLE=iio_trigger_example ADXL355=y"
     }
   }
 }
