+++ projects/eval-adxl355-pmdz/src.mk
@@ -1,16 +1,3 @@
-include $(PROJECT)/src/platform/$(PLATFORM)/platform_src.mk
-include $(PROJECT)/src/examples/examples_src.mk
-
-SRCS += $(PROJECT)/src/platform/$(PLATFORM)/main.c
-
-INCS += $(PROJECT)/src/common/common_data.h
-SRCS += $(PROJECT)/src/common/common_data.c
-	
-INCS += $(PROJECT)/src/platform/platform_includes.h
-
-INCS += $(PROJECT)/src/platform/$(PLATFORM)/parameters.h
-SRCS += $(PROJECT)/src/platform/$(PLATFORM)/parameters.c 
-
 INCS += $(INCLUDE)/no_os_delay.h     \
 		$(INCLUDE)/no_os_error.h     \
 		$(INCLUDE)/no_os_gpio.h      \
@@ -44,12 +31,3 @@ SRCS += $(DRIVERS)/api/no_os_gpio.c \
 
 INCS += $(DRIVERS)/accel/adxl355/adxl355.h
 SRCS += $(DRIVERS)/accel/adxl355/adxl355.c
-
-ifdef IIO_LWIP_EXAMPLE
-INCS += $(INCLUDE)/no_os_crc8.h
-INCS += $(DRIVERS)/net/adin1110/adin1110.h
-INCS += $(NO-OS)/network/lwip_raw_socket/netdevs/adin1110/lwip_adin1110.h
-SRCS += $(NO-OS)/network/lwip_raw_socket/netdevs/adin1110/lwip_adin1110.c
-SRCS += $(DRIVERS)/net/adin1110/adin1110.c
-SRCS += $(NO-OS)/util/no_os_crc8.c
-endif
