+++ projects/eval-adxl355-pmdz/src/common/common_data.h
@@ -36,17 +36,8 @@
 /******************************************************************************/
 /***************************** Include Files **********************************/
 /******************************************************************************/
-#include "platform_includes.h"
+#include "parameters.h"
 #include "adxl355.h"
-#ifdef IIO_SUPPORT
-#include "iio_adxl355.h"
-#ifdef IIO_TRIGGER_EXAMPLE
-#include "iio_trigger.h"
-#endif
-#ifdef IIO_LWIP_EXAMPLE
-#include "adin1110.h"
-#endif
-#endif
 
 /******************************************************************************/
 /********************** Macros and Constants Definitions **********************/
@@ -54,17 +45,5 @@
 extern struct no_os_uart_init_param adxl355_uart_ip;
 extern struct no_os_spi_init_param adxl355_spi_ip;
 extern struct adxl355_init_param adxl355_ip;
-
-#ifdef IIO_TRIGGER_EXAMPLE
-#define ADXL355_GPIO_TRIG_NAME "adxl355-dev0"
-
-extern struct iio_hw_trig_init_param adxl355_gpio_trig_ip;
-extern struct no_os_irq_init_param adxl355_gpio_irq_ip;
-#endif
-
-#ifdef IIO_LWIP_EXAMPLE
-extern uint8_t adin1110_mac_address[6];
-extern struct adin1110_init_param adin1110_ip;
-#endif
 
 #endif /* __COMMON_DATA_H__ */
