+++ projects/eval-adxl355-pmdz/src/examples/iio_example/iio_example.mk
@@ -1,0 +1,3 @@
+IIOD = y
+INCS += $(DRIVERS)/accel/adxl355/iio_adxl355.h
+SRCS += $(DRIVERS)/accel/adxl355/iio_adxl355.c
