+++ projects/eval-adxl355-pmdz/src/examples/iio_lwip_example/iio_lwip_example.c
@@ -34,12 +34,12 @@
 /******************************************************************************/
 /***************************** Include Files **********************************/
 /******************************************************************************/
-#include "iio_lwip_example.h"
 #include "iio_adxl355.h"
 #include "common_data.h"
 #include "iio_app.h"
 #include "lwip_socket.h"
 #include "lwip_adin1110.h"
+#include "adin1110.h"
 
 /******************************************************************************/
 /********************** Macros and Constants Definitions **********************/
@@ -53,6 +53,33 @@
 /******************************************************************************/
 uint8_t iio_data_buffer[DATA_BUFFER_SIZE * 3 * sizeof(int)];
 
+uint8_t adin1110_mac_address[6] = {0x00, 0x18, 0x80, 0x03, 0x25, 0x80};
+
+const struct no_os_gpio_init_param adin1110_reset_ip = {
+	.port = ADIN1110_GPIO_RESET_PORT,
+	.number = ADIN1110_GPIO_RESET_PIN,
+	.pull = NO_OS_PULL_NONE,
+	.platform_ops = GPIO_OPS,
+	.extra = &adin1110_reset_gpio_extra_ip,
+};
+
+const struct no_os_spi_init_param adin1110_spi_ip = {
+	.device_id = ADIN1110_SPI_DEVICE_ID,
+	.max_speed_hz = ADIN1110_SPI_CLK_RATE,
+	.bit_order = NO_OS_SPI_BIT_ORDER_MSB_FIRST,
+	.mode = NO_OS_SPI_MODE_0,
+	.platform_ops = SPI_OPS,
+	.chip_select = ADIN1110_SPI_CS,
+	.extra = &adin1110_spi_extra_ip,
+};
+
+struct adin1110_init_param adin1110_ip = {
+	.chip_type = ADIN1110,
+	.comm_param = adin1110_spi_ip,
+	.reset_param = adin1110_reset_ip,
+	.append_crc = false,
+};
+
 /******************************************************************************/
 /************************ Functions Definitions *******************************/
 /******************************************************************************/
@@ -63,7 +90,7 @@ uint8_t iio_data_buffer[DATA_BUFFER_SIZE * 3 * sizeof(int)];
  *               execute continuously function iio_app_run and will not return.
 *******************************************************************************/
 
-int iio_lwip_example_main()
+int example_main()
 {
 	int ret;
 	struct adxl355_iio_dev *adxl355_iio_desc;
