+++ projects/eval-adxl355-pmdz/src/platform/aducm3029/main.c
@@ -34,20 +34,10 @@
 /******************************************************************************/
 /***************************** Include Files **********************************/
 /******************************************************************************/
-#include "platform_includes.h"
+#include "parameters.h"
 #include "common_data.h"
 
-#ifdef DUMMY_EXAMPLE
-#include "dummy_example.h"
-#endif
-
-#ifdef IIO_EXAMPLE
-#include "iio_example.h"
-#endif
-
-#ifdef IIO_TRIGGER_EXAMPLE
-#include "iio_trigger_example.h"
-#endif
+extern int example_main();
 
 /***************************************************************************//**
  * @brief Main function execution for ADUCM3029 platform.
@@ -57,62 +47,19 @@
 int main()
 {
 	int ret;
-	adxl355_ip.comm_init.spi_init = adxl355_spi_ip;
-
-	ret = platform_init();
-	if (ret)
-		goto error;
-
-#ifdef DUMMY_EXAMPLE
-	struct no_os_uart_desc *uart_desc;
-
-	ret = no_os_uart_init(&uart_desc, &adxl355_uart_ip);
-	if (ret)
-		goto error;
-
-	no_os_uart_stdio(uart_desc);
-
-	ret = dummy_example_main();
-	if (ret)
-		goto error_uart;
-error_uart:
-	return no_os_uart_remove(uart_desc);
-#endif
-
-#ifdef IIO_EXAMPLE
-	ret = iio_example_main();
-	if (ret)
-		goto error;
-#endif
-
-#ifdef IIO_TRIGGER_EXAMPLE
-
 	struct no_os_gpio_desc *adxl355_gpio_desc;
 
-	/* Initialize DATA READY pin */
+	adxl355_ip.comm_init.spi_init = adxl355_spi_ip;
+
+	platform_init();
+
 	ret = no_os_gpio_get_optional(&adxl355_gpio_desc, &adxl355_gpio_drdy_ip);
 	if (ret)
-		goto error;
+		return ret;
 
 	ret = no_os_gpio_direction_input(adxl355_gpio_desc);
 	if (ret)
-		goto error_gpio;
+		return ret;
 
-	ret = iio_trigger_example_main();
-	if (ret)
-		goto error_gpio;
-
-error_gpio:
-	return no_os_gpio_remove(adxl355_gpio_desc);
-#endif
-
-#if (DUMMY_EXAMPLE + IIO_EXAMPLE + IIO_TRIGGER_EXAMPLE == 0)
-#error At least one example has to be selected using y value in Makefile.
-#elif (DUMMY_EXAMPLE + IIO_EXAMPLE + IIO_TRIGGER_EXAMPLE > 1)
-#error Selected example projects cannot be enabled at the same time. \
-Please enable only one example and rebuild the project.
-#endif
-
-error:
-	return ret;
+	return example_main();
 }
