+++ projects/eval-adxl355-pmdz/src/platform/maxim/main.c
@@ -34,21 +34,11 @@
 /******************************************************************************/
 /***************************** Include Files **********************************/
 /******************************************************************************/
-#include "platform_includes.h"
+#include "parameters.h"
 #include "common_data.h"
 #include "no_os_error.h"
 
-#ifdef IIO_EXAMPLE
-#include "iio_example.h"
-#endif
-
-#ifdef IIO_TRIGGER_EXAMPLE
-#include "iio_trigger_example.h"
-#endif
-
-#ifdef DUMMY_EXAMPLE
-#include "dummy_example.h"
-#endif
+extern int example_main();
 
 /***************************************************************************//**
  * @brief Main function execution for Maxim platform.
@@ -61,16 +51,8 @@ int main()
 
 	adxl355_ip.comm_init.spi_init = adxl355_spi_ip;
 
-#ifdef IIO_EXAMPLE
-	ret = iio_example_main();
-#endif
-
-#ifdef IIO_TRIGGER_EXAMPLE
+#ifdef GPIO_DRDY_PORT_NUM
 	struct no_os_gpio_desc *adxl355_gpio_desc;
-	struct no_os_irq_ctrl_desc *nvic_desc;
-	struct no_os_irq_init_param nvic_ip = {
-		.platform_ops = &max_irq_ops,
-	};
 
 	/* Initialize DATA READY pin */
 	ret = no_os_gpio_get_optional(&adxl355_gpio_desc, &adxl355_gpio_drdy_ip);
@@ -80,7 +62,12 @@ int main()
 	ret = no_os_gpio_direction_input(adxl355_gpio_desc);
 	if (ret)
 		return ret;
-
+#endif
+#ifdef NVIC_GPIO_IRQ
+	struct no_os_irq_ctrl_desc *nvic_desc;
+	struct no_os_irq_init_param nvic_ip = {
+		.platform_ops = &max_irq_ops,
+	};
 	/* Initialize GPIO IRQ controller */
 	ret = no_os_irq_ctrl_init(&nvic_desc, &nvic_ip);
 	if (ret)
@@ -89,28 +76,8 @@ int main()
 	ret = no_os_irq_enable(nvic_desc, NVIC_GPIO_IRQ);
 	if (ret)
 		return ret;
-
-	ret = iio_trigger_example_main();
 #endif
 
-#ifdef DUMMY_EXAMPLE
-	struct no_os_uart_desc *uart_desc;
-
-	ret = no_os_uart_init(&uart_desc, &adxl355_uart_ip);
-	if (ret)
-		return ret;
-
-	no_os_uart_stdio(uart_desc);
-	ret = dummy_example_main();
-#endif
-
-#if (DUMMY_EXAMPLE + IIO_EXAMPLE + IIO_TRIGGER_EXAMPLE == 0)
-#error At least one example has to be selected using y value in Makefile.
-#elif (DUMMY_EXAMPLE + IIO_EXAMPLE + IIO_TRIGGER_EXAMPLE > 1)
-#error Selected example projects cannot be enabled at the same time. \
-Please enable only one example and rebuild the project.
-#endif
-
-	return ret;
+	return example_main();
 }
