+++ projects/eval-adxl355-pmdz/src/platform/stm32/parameters.c
@@ -47,13 +47,10 @@ struct stm32_spi_init_param adxl355_spi_extra_ip  = {
 	.chip_select_port = SPI_CS_PORT,
 };
 
-#ifdef IIO_TRIGGER_EXAMPLE
 struct stm32_gpio_irq_init_param adxl355_gpio_irq_extra_ip = {
 	.port_nb = 0, /* Port A */
 };
-#endif
 
-#ifdef IIO_LWIP_EXAMPLE
 struct stm32_gpio_init_param adin1110_reset_gpio_extra_ip = {
 	.mode = GPIO_MODE_OUTPUT_OD,
 	.speed = GPIO_SPEED_FREQ_VERY_HIGH,
@@ -62,4 +59,3 @@ struct stm32_gpio_init_param adin1110_reset_gpio_extra_ip = {
 struct stm32_spi_init_param adin1110_spi_extra_ip  = {
 	.chip_select_port = ADIN1110_SPI_CS_PORT,
 };
-#endif
